Move get_force_get_task_result to api_testing_common.py
